requirements.txt
python-telegram-bot==21.7
notion-client==2.2.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
boto3==1.39.9
botocore==1.39.9 